backend/requirements.txt
flask
flask-cors
gunicorn
pandas
scikit-learn
pillow
PyPDF2
python-docx
requests